+++ inherited-rules/implicitly-included/eslint_conf_eslint-recommended.js
@@ -2,7 +2,6 @@ module.exports = {
   rules: {
     "for-direction": "error",
     "getter-return": "error",
-    "no-case-declarations": "error",
     "no-extra-semi": "error",
     "no-unused-labels": "error",
     "require-atomic-updates": "error",