+++ CHANGES.md
@@ -2,7 +2,7 @@ This changelog follows Semantic Versioning https://semver.org/
 
 # 2.0.0 (2019-07-24)
 
-### Major 
+### Major
 
 * Updated from upstream, JSDoc changes will make current packages fail and will need minor changes